pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csfl-simulator"
version = "0.1.0"
description = "Streamlit-based federated client selection simulator"
authors = [{name = "CSFL Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "streamlit",
  "streamlit-ace",
  "torch",
  "torchvision",
  "numpy",
  "pandas",
  "scikit-learn",
  "matplotlib",
  "plotly",
  "seaborn",
  "nbformat",
  "nbclient",
  "pyyaml",
  "rich",
  "tqdm",
  "networkx",
  "thop"
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools.packages.find]
include = ["csfl_simulator*"]
